grabs a specified SpongeBob character lines
This is the initial script that is plenty useful by itself. I plan to update it so it can get the previous line of dialogue to create prompt and response pairs. Not sure if I'm going to add and change this repo from a parser to the full project or just create another repo.